refactor: extract editor panel HTML to template module [swarm]
